efficientnet/scripts/requirements.txt
tensorflowjs==1.1.2
tf-nightly-2.0-preview==2.0.0.dev20190623
